Fix int to float warnings

+++ Engine/Dude.h
@@ -6,7 +6,7 @@ class Dude {
 
 public:
 
-	Dude(int in_x, int in_y)
+	Dude(float in_x, float in_y)
 	{
 		x = in_x;
 		y = in_y;
